pyproject.toml
[project]
name = "goldilocks-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "httpx>=0.28",
    "python-dotenv>=1.0",
    "goldilocks-core",
    "mp-api>=0.41",
]

[tool.uv.sources]
goldilocks-core = { path = "../goldilocks-core", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
